Temporary commit, need to fix relese.

+++ src/tests/utils/socket.c
@@ -817,7 +817,7 @@ util_err_t socket_sendtov_buf(socket_t sock, socket_address_t *addr, buf_t **seg
     msg.msg_name = (void*)&addr->addr;
     msg.msg_namelen = addr->addr_len;
     msg.msg_iov = vecs;
-    msg.msg_iovlen = (size_t)vec_count;
+    msg.msg_iovlen = (int)vec_count;
 
 #ifdef UTIL_BSD_OS_TYPE
     /* On BSD/macOS, SO_NOSIGPIPE was set at socket creation */
